# fix: use @npmcli/redact for log redactions