--- a/workspaces/arborist/package.json
+++ b/workspaces/arborist/package.json
@@ -24,7 +24,7 @@
     "npm-install-checks": "^6.2.0",
     "npm-package-arg": "^11.0.1",
     "npm-pick-manifest": "^9.0.0",
-    "npm-registry-fetch": "^16.0.0",
+    "npm-registry-fetch": "^16.2.0",
     "npmlog": "^7.0.1",
     "pacote": "^17.0.4",
     "parse-conflict-json": "^3.0.0",