--- a/workspaces/libnpmpublish/package.json
+++ b/workspaces/libnpmpublish/package.json
@@ -41,7 +41,7 @@
     "ci-info": "^4.0.0",
     "normalize-package-data": "^6.0.0",
     "npm-package-arg": "^11.0.1",
-    "npm-registry-fetch": "^16.0.0",
+    "npm-registry-fetch": "^16.2.0",
     "proc-log": "^3.0.0",
     "semver": "^7.3.7",
     "sigstore": "^2.2.0",