… lib/npm.js
… lib/utils/replace-info.js
… test/lib/utils/replace-info.js