Closes #7314

This does the same thing as #7314 except using a new shared package for the log cleaning `@npmcli/redact`.

Here are the perf numbers from my local machine after setting a script with `npm pkg set scripts.echo="echo"`. Note that the before seems slower on my machine than previous benchmarks, I think due to some security software that affects filesystem IO performance.

**Before**

```sh
❯ hyperfine --warmup 3 "node ./bin/npm-cli.js run echo"
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     349.4 ms ±  20.2 ms    [User: 172.5 ms, System: 42.7 ms]
  Range (min … max):   319.0 ms … 376.1 ms    10 runs
```

**After**

```sh
❯ hyperfine --warmup 3 "node ./bin/npm-cli.js run echo"
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     170.7 ms ±   4.7 ms    [User: 114.7 ms, System: 22.6 ms]
  Range (min … max):   163.6 ms … 184.9 ms    16 runs
```