@H4ad 2024-04-03
Before:

```
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     232.8 ms ±   3.2 ms    [User: 235.4 ms, System: 54.8 ms]
  Range (min … max):   226.4 ms … 237.3 ms    12 runs
```

After:

```
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     196.0 ms ±   2.5 ms    [User: 179.1 ms, System: 53.8 ms]
  Range (min … max):   192.0 ms … 199.5 ms    14 runs
```

The new load time for each dependency (top N of most expensive):

```
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/semver/functions/satisfies.js] [../classes/range]: 6.436ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/validate-npm-package-name/lib/index.js] [builtins]: 6.441ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/utils/error-message.js] [validate-npm-package-name]: 6.657ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/cli-entry.js] [semver/functions/satisfies]: 6.811ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/npmlog/lib/log.js] [gauge]: 7.489ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/normalize-package-data/lib/fixer.js] [validate-npm-package-license]: 7.668ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/utils/exit-handler.js] [./error-message.js]: 7.684ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/utils/log-shim.js] [npmlog]: 9.684ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/utils/exit-handler.js] [./log-shim.js]: 10.266ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/workspaces/config/lib/index.js] [@npmcli/map-workspaces]: 10.349ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/@npmcli/package-json/lib/normalize.js] [normalize-package-data/lib/fixer.js]: 10.578ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/npm.js] [@npmcli/config]: 15.221ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/@npmcli/package-json/lib/index.js] [./normalize.js]: 17.558ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/node_modules/@npmcli/run-script/lib/run-script.js] [@npmcli/package-json]: 18.414ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/cli-entry.js] [./utils/exit-handler.js]: 18.496ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/commands/run-script.js] [@npmcli/run-script]: 20.343ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/npm.js] [./commands/run-script.js]: 22.623ms
MODULE_CJS 60649 [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/lib/cli-entry.js] [./npm.js]: 25.942ms
MODULE_CJS 60649 [] [/home/h4ad/Projects/opensource/performance-analysis/npm-cli/bin/npm-cli.js]: 60.411ms
```